6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int main(void)
 	long int w;
 	long long int u;
 	float z;
-printf("Size of a char: %lu byte(s)\n", sizeof(x));
-printf("size of int: %lu byte(s)\n", sizeof(y));
-printf("size of long int: %lu byte(s)\n"' sizeof(w));
-printf("size of long long int: %lu byte(s)\n", sizeof(u));
-printf("size of fload: %lu byte(s)\n", sizeof(z))
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(x));
+printf("size of int: %lu byte(s)\n", (unsigned long)sizeof(y));
+printf("size of long int: %lu byte(s)\n"' (unsigned long)sizeof(w));
+printf("size of long long int: %lu byte(s)\n", (unsigned long)sizeof(u));
+printf("size of fload: %lu byte(s)\n", (unsigned long)sizeof(z))
 return (0);
 }
